requirements.txt
absl-py==2.3.0
aiomysql==0.2.0
anyio==4.6.2
astunparse==1.6.3
bcrypt==4.0.1
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cryptography==42.0.5
cycler==0.12.1
deepface==0.0.93
dnspython==2.6.1
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.110.0
filelock==3.18.0
fire==0.7.0
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.2.10
fonttools==4.58.4
fsspec==2025.5.1
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
greenlet==3.2.3
grpcio==1.73.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.14.0
idna==3.7
inflect==7.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
keras==3.10.0
keras-facenet==0.3.2
kiwisolver==1.4.8
libclang==18.1.1
lz4==4.4.4
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.1
more-itertools==10.5.0
mpmath==1.3.0
mtcnn==1.0.0
multidict==6.0.5
mypy_extensions==1.1.0
namex==0.1.0
networkx==3.4.2
numpy==1.26.4
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.3.0
passlib==1.7.4
Pillow==10.1.0
propcache==0.2.0
protobuf==5.29.5
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.5.1
pycparser==2.22
pydantic==1.10.15
pydantic_core==2.16.3
Pygments==2.19.1
PyMySQL==1.1.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.1
requests==2.32.4
retina-face==0.0.17
rich==14.0.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlacodegen==3.0.0
SQLAlchemy==2.0.29
starlette==0.36.3
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
tf_keras==2.19.0
threadpoolctl==3.6.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typeguard==4.2.1
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2025.2
ultralytics==8.3.155
ultralytics-thop==2.0.14
urllib3==2.4.0
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.9.4
